% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarNetwork.R
\name{multivarNetwork}
\alias{multivarNetwork}
\title{Fit a collection of networks from multivariate experiments}
\usage{
multivarNetwork(X, sym.rule = "AND", cv.choice = c("none", "1se", "min"),
  nlambda = 50, min.ratio = 0.001, mc.cores = 1)
}
\arguments{
\item{X}{a list of matrices with the same dimension, corresponding to multiple experiments related
to the same individuals and variables, e.g., proteomics and transcriptomics data.}

\item{sym.rule}{a character, either "AND" or "OR" to define the  post-symmetrization rule applied to the coefficients in neighborhood selection}

\item{cv.choice}{a character for defining the cross-validation rule used to choose the penalty level (i.e., the number of edges in the the network).
If "none", the whole regulariazation paths is sent back. If "min" or "1se", penalties are cross-validation for each variable
with the corresponding rule.}

\item{nlambda}{integer defining the number of penalty levels used on the grid}

\item{min.ratio}{the grid of penalties starts by lambda.max, that is,  minimal penalty level for selecting no edge in the network, then decrease on a
log scale until min.ratio* lambda.max.}

\item{mc.cores}{for distributing the computation over the variables. Default is 1 (no parallelization).}
}
\description{
The inference procedure is multivariate neighborhood selection with group Lasso penalty.
If an univariate experiment is provided, the usual neighborhood selection with lasso penalty applies.
}
